Support for externally provided decompressed supplements

+++ panoramix/utils/supplement.py
@@ -1,8 +1,10 @@
 import json
 import logging
 import lzma
+import shutil
 from pathlib import Path
 
+from panoramix.env import DECOMPRESSED_SUPPLEMENTS
 from panoramix.utils.db import get_sqlite3_cursor
 from panoramix.utils.helpers import cache_dir, cached
 
@@ -42,11 +44,15 @@ def supplements_path():
 def check_supplements():
     panoramix_supplements = supplements_path()
     if not panoramix_supplements.is_file():
-        compressed_supplements = Path(__file__).parent.parent / "data" / "supplement.db.xz"
-        logger.info("Decompressing %s into %s...", compressed_supplements, panoramix_supplements)
-        with lzma.open(compressed_supplements) as inf, panoramix_supplements.open("wb") as outf:
-            while buf := inf.read(1024 * 1024):
-                outf.write(buf)
+        if DECOMPRESSED_SUPPLEMENTS and Path(DECOMPRESSED_SUPPLEMENTS).is_file():
+            logger.info("Copying %s into %s...", DECOMPRESSED_SUPPLEMENTS, panoramix_supplements)
+            shutil.copy(DECOMPRESSED_SUPPLEMENTS, panoramix_supplements)
+        else:
+            compressed_supplements = Path(__file__).parent.parent / "data" / "supplement.db.xz"
+            logger.info("Decompressing %s into %s...", compressed_supplements, panoramix_supplements)
+            with lzma.open(compressed_supplements) as inf, panoramix_supplements.open("wb") as outf:
+                while buf := inf.read(1024 * 1024):
+                    outf.write(buf)
 
     assert panoramix_supplements.is_file()
